+++ xiao_lvgl.c
@@ -1,0 +1,59 @@
+
+#include "esp_log.h"
+#include "esp_check.h"
+#include "esp_lvgl_port.h"
+#include "esp_lcd_touch.h"
+#include "esp_lcd_types.h"
+
+#include "xiao_common.h"
+#include "xiao_lvgl.h"
+
+static lv_disp_t *lvgl_disp = NULL;
+extern esp_lcd_panel_io_handle_t io;
+extern esp_lcd_panel_handle_t panel;
+extern esp_lcd_touch_handle_t handle;
+
+esp_err_t init_lvgl()
+{
+    /* Initialize LVGL */
+    const lvgl_port_cfg_t lvgl_cfg = {
+        .task_priority = 4,         /* LVGL task priority */
+        .task_stack = 4096,         /* LVGL task stack size */
+        .task_affinity = -1,        /* LVGL task pinned to core (-1 is no affinity) */
+        .task_max_sleep_ms = 500,   /* Maximum sleep in LVGL task */
+        .timer_period_ms = 5        /* LVGL timer tick period in ms */
+    };
+
+    ESP_RETURN_ON_ERROR(lvgl_port_init(&lvgl_cfg), TAG, "LVGL port initialization failed");
+
+    /* Add LCD screen */
+    ESP_LOGD(TAG, "Add LCD screen");
+    const lvgl_port_display_cfg_t disp_cfg = {
+        .io_handle = io,
+        .panel_handle = panel,
+        .buffer_size = XIAO_BUFFER_SIZE,
+        .double_buffer = 1,
+        .hres = XIAO_HOR_RES_MAX,
+        .vres = XIAO_VIR_RES_MAX,
+        .monochrome = false,
+        /* Rotation values must be same as used in esp_lcd for initial settings of the screen */
+        .rotation = {
+            .swap_xy = false,
+            .mirror_x = true,
+            .mirror_y = true,
+        },
+        .flags = {
+            .buff_dma = true,
+        }
+    };
+    lvgl_disp = lvgl_port_add_disp(&disp_cfg);
+
+    /* Add touch input (for selected screen) */
+    const lvgl_port_touch_cfg_t touch_cfg = {
+        .disp = lvgl_disp,
+        .handle = handle,
+    };
+    lvgl_port_add_touch(&touch_cfg);
+
+    return ESP_OK;
+}