+++ xiao_sd_card.c
@@ -1,0 +1,55 @@
+
+#include "esp_log.h"
+#include "esp_check.h"
+#include "driver/gpio.h"
+#include "driver/spi_master.h"
+#include "driver/sdmmc_host.h"
+#include "esp_vfs_fat.h"
+#include "sdmmc_cmd.h"
+
+#include "xiao_common.h"
+#include "xiao_sd_card.h"
+
+/* SD card */
+sdmmc_card_t *card = NULL;
+
+esp_err_t xiao_round_display_sd_card_init()
+{
+    esp_err_t ret;
+    sdmmc_host_t host = SDSPI_HOST_DEFAULT();
+
+    esp_vfs_fat_sdmmc_mount_config_t mount_config = {
+        .format_if_mount_failed = false,
+        .max_files = 5,
+        .allocation_unit_size = 16 * 1024
+    };
+
+    sdspi_device_config_t slot_config = SDSPI_DEVICE_CONFIG_DEFAULT();
+    slot_config.gpio_cs = XIAO_SD_CS;
+    slot_config.host_id = host.slot;
+
+    ESP_ERROR_CHECK(gpio_set_level(XIAO_LCD_CS, GPIO_LEVEL_HIGH));
+    ESP_LOGI(TAG, "Mounting filesystem");
+    ret = esp_vfs_fat_sdspi_mount(MOUNT_POINT, &host, &slot_config, &mount_config, &card);
+
+    if (ret != ESP_OK) {
+        if (ret == ESP_FAIL) {
+            ESP_LOGE(TAG, "Failed to mount filesystem. "
+                     "If you want the card to be formatted, set the CONFIG_EXAMPLE_FORMAT_IF_MOUNT_FAILED menuconfig option.");
+        } else {
+            ESP_LOGE(TAG, "Failed to initialize the card (%s). "
+                     "Make sure SD card lines have pull-up resistors in place.", esp_err_to_name(ret));
+        }
+        return ret;
+    }
+    ESP_LOGI(TAG, "Filesystem mounted");
+
+    return ret;
+}
+
+void xiao_round_display_sd_card_deinit()
+{
+    // All done, unmount partition and disable SPI peripheral
+    esp_vfs_fat_sdcard_unmount(MOUNT_POINT, card);
+    ESP_LOGI(TAG, "Card unmounted");
+}