+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.1.0] - 2023-11-21
+
+- Add custom cst816s driver for xiao round display
+- Add custom esp_lvgl_port which includes touch interupt
+
+
 ## [1.0.1] - 2023-11-21
 
 - Fix upload component workflow
@@ -14,3 +20,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - First release
 
+## [2.0.0] - 2023-12-24
+- Fix i2c confliction